+++ blocks/form/index.asset.php
@@ -7,5 +7,5 @@
 
 return array(
 	'dependencies' => array( 'wp-blocks', 'wp-element', 'wp-block-editor', 'wp-components', 'wp-i18n', 'wp-server-side-render', 'wp-api-fetch' ),
-	'version'      => '1.3.0',
+	'version'      => '1.4.0',
 );